example-clients/python/requirements.txt
requests
typing_extensions
google-auth